// enums.ts
export enum EmergencyType {
    SEVERE_BLEEDING = "severe_bleeding",
    CAR_ACCIDENT = "car_accident",
    MOTORCYCLE_ACCIDENT = "motorcycle_accident",
    BUILDING_COLLAPSE = "building_collapse",
    FALL_FROM_HEIGHT = "fall_from_height",
    GUNSHOT_WOUND = "gunshot_wound",
    STABBING_OR_MACHETE = "stabbing_or_machete",
    BURN_INJURY = "burn_injury",
    ELECTRIC_SHOCK = "electric_shock",
    DROWNING = "drowning",
    SNAKE_BITE = "snake_bite",
    ANIMAL_BITE = "animal_bite",
    POISONING = "poisoning",
    DRUG_OVERDOSE = "drug_overdose",
    HEART_ATTACK = "heart_attack",
    CHEST_PAIN = "chest_pain",
    STROKE_SYMPTOMS = "stroke_symptoms",
    UNCONSCIOUS = "unconscious",
    SEIZURE = "seizure",
    DIFFICULTY_BREATHING = "difficulty_breathing",
    ASTHMA_ATTACK = "asthma_attack",
    SEVERE_ALLERGIC_REACTION = "severe_allergic_reaction",
    BROKEN_BONE = "broken_bone",
    HEAD_INJURY = "head_injury",
    SPINAL_INJURY = "spinal_injury",
    PREGNANCY_EMERGENCY = "pregnancy_emergency",
    CHILD_EMERGENCY = "child_emergency",
    HIGH_FEVER_CHILD = "high_fever_child",
    UNKNOWN_EMERGENCY = "unknown_emergency"
}